fastvep-sa: address Copilot review
- block.rs: stream-decompress with Read::take() so MAX_BLOCK_DECOMPRESSED
  is enforced *during* decompression. zstd::decode_all materializes the
  full output before any size check, leaving the reader vulnerable to
  zstd compression bombs.

- common.rs: store MAX_INDEX_PAYLOAD as u64 (was usize). The literal
  4 * 1024 * 1024 * 1024 == 2^32 doesn't fit in 32-bit usize. Update
  index/gene/interval read_from to compare in u64 and TryFrom into
  usize before allocating.

- reader.rs preload: compute (min, max) in a single fold pass instead
  of two separate iterator scans. Comment was misleading.

+++ crates/fastvep-sa/src/common.rs
@@ -21,7 +21,11 @@ pub const ZSTD_LEVEL: i32 = 3;
 /// Hard cap on a single bincode-serialized index payload (4 GiB). Used by
 /// `.osa.idx`, `.osi`, and `.oga` readers to refuse malformed/malicious files
 /// that claim absurd payload sizes, before allocating a buffer.
-pub const MAX_INDEX_PAYLOAD: usize = 4 * 1024 * 1024 * 1024;
+///
+/// Stored as `u64` so the literal compiles on 32-bit targets (where
+/// `usize` is 32 bits and cannot hold `2^32`). Readers must additionally
+/// verify the value fits in `usize` before allocating.
+pub const MAX_INDEX_PAYLOAD: u64 = 4 * 1024 * 1024 * 1024;
 
 /// File extension for position/allele-level annotations.
 pub const OSA_EXT: &str = "osa";
